Updated processes + video endpoint
