+++ src/app.py
@@ -1,49 +1,28 @@
 from vehicle.consts import LIDAR_USB_HEADER, LIDAR_DATA_PATH
 from joystick import JOYSTICK_BP
 from routes import ROUTES_BP
-from car import Robot, VEHICLE
+from car import Robot
 from multiprocessing import Process
-from flask import Flask, Response
-import cv2
+from flask import Flask
+
+APP_EXECUTION_TIME = 600 # seconds
 
 app = Flask(__name__)
 app.register_blueprint(JOYSTICK_BP)
 app.register_blueprint(ROUTES_BP)
 
-@app.route('/video_feed', methods=['GET'])
-def video_feed():
-    """
-        Endpoint function that continuously returns a captured image from the camera. 
-    """
-
-    def generate():
-        camera = cv2.VideoCapture(0)
-        while True:
-                success, frame = camera.read()
-                if (not success):
-                     break
-                else:
-                #frame = VEHICLE.camera().capture_array()
-                    ret, buffer = cv2.imencode('.jpg', frame)
-                    frame = buffer.tobytes()
-                    yield (b'--frame\r\n'
-                            b'Content-Type: image/jpeg\r\n\r\n' + frame + b'\r\n')
-    return Response(generate(),
-                    mimetype='multipart/x-mixed-replace; boundary=frame')
-
 if (__name__ == '__main__'):
     flask_process = Process(target=app.run)
     flask_process.start()
 
-    #robot = Robot(LIDAR_DATA_PATH, LIDAR_USB_HEADER)
-    #robot_process = Process(target=robot.run)
-    #robot_process.start()
+    robot = Robot(LIDAR_DATA_PATH, LIDAR_USB_HEADER)
+    robot_process = Process(target=robot.run)
+    robot_process.start()
 
-    #robot_process.join(15)
-    flask_process.join(15)
+    robot_process.join(APP_EXECUTION_TIME)
+    flask_process.join(1)
     
-    #robot_process.terminate()
+    robot_process.terminate()
     flask_process.terminate()
 
-    #robot.stop_lidar()
-    #VEHICLE.camera().stop()
+    robot.stop_lidar()
