+++ src/routes.py
@@ -13,6 +13,26 @@ def index():
     VEHICLE.set_manual_mode()
     return render_template('index.html')
 
+@ROUTES_BP.route('/video_feed', methods=['GET'])
+def video_feed():
+    """
+        Endpoint function that continuously returns a captured image from the camera. 
+    """
+
+    def generate():
+        camera = cv2.VideoCapture(0)
+        while True:
+                success, frame = camera.read()
+                if (not success):
+                     break
+                else:
+                    ret, buffer = cv2.imencode('.jpg', frame)
+                    frame = buffer.tobytes()
+                    yield (b'--frame\r\n'
+                            b'Content-Type: image/jpeg\r\n\r\n' + frame + b'\r\n')
+    return Response(generate(),
+                    mimetype='multipart/x-mixed-replace; boundary=frame')
+
 @ROUTES_BP.route('/autonomous-mode', methods=['POST'])
 def setAutonomousMode():
     """
